Refactor System.IO.Pipes with mini-PAL
This commit separates out the Win32 code used by System.IO.Pipes into its own files; these files are then conditionally included based on the OS property supplied to MSBuild.  Files for Unix are included as well, with stubbed out implementations.
