+++ src/System.IO.Pipes/src/Interop/Interop.Windows.cs
