+++ src/System.IO.Pipes/src/Properties/AssemblyInfo.cs
@@ -1,6 +1,2 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
