+++ src/System.IO.Pipes/src/System.IO.Pipes.csproj
@@ -15,41 +15,64 @@
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Release|AnyCPU'">
   </PropertyGroup>
+
   <!-- References are resolved from packages.config -->
   <!-- Compiled Source Files -->
   <ItemGroup>
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Resources\Strings.Designer.cs">
-      <AutoGen>True</AutoGen>
-      <DesignTime>True</DesignTime>
-      <DependentUpon>Strings.resx</DependentUpon>
+    <Compile Include="..\..\Common\src\System\SR.cs">
+      <Link>System\SR.cs</Link>
     </Compile>
-    <Compile Include="Interop\Interop.Manual.cs" />
-    <Compile Include="System\IO\Pipes\IOCancellationHelper.cs" />
-    <Compile Include="System\IO\Pipes\Enums.cs" />
-    <Compile Include="System\IO\Pipes\Pipe.cs" />
-    <Compile Include="System\IO\Pipes\PipeStream.cs" />
-    <Compile Include="Microsoft\Win32\SafeHandles\SafePipeHandle.cs" />
-    <Compile Include="System\IO\__Error.cs" />
-  </ItemGroup>
-  <!-- Common files -->
-  <ItemGroup>
     <Compile Include="..\..\Common\src\System\IO\StreamAsyncHelper.cs">
       <Link>System\IO\StreamAsyncHelper.cs</Link>
     </Compile>
     <Compile Include="..\..\Common\src\System\IO\TaskHelpers.cs">
       <Link>System\IO\TaskHelpers.cs</Link>
     </Compile>
+    <Compile Include="Microsoft\Win32\SafeHandles\SafePipeHandle.cs" />
+    <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Resources\Strings.Designer.cs">
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+      <DependentUpon>Strings.resx</DependentUpon>
+    </Compile>
+    <Compile Include="System\IO\__Error.cs" />
+    <Compile Include="System\IO\Pipes\AnonymousPipeClientStream.cs" />
+    <Compile Include="System\IO\Pipes\AnonymousPipeServerStream.cs" />
+    <Compile Include="System\IO\Pipes\PipeDirection.cs" />
+    <Compile Include="System\IO\Pipes\PipeOptions.cs" />
+    <Compile Include="System\IO\Pipes\PipeTransmissionMode.cs" />
+    <Compile Include="System\IO\Pipes\NamedPipeClientStream.cs" />
+    <Compile Include="System\IO\Pipes\NamedPipeServerStream.cs" />
+    <Compile Include="System\IO\Pipes\PipeState.cs" />
+    <Compile Include="System\IO\Pipes\PipeStream.cs" />
+  </ItemGroup>
+
+  <ItemGroup Condition=" '$(OS)' == 'Windows_NT' ">
     <Compile Include="..\..\Common\src\System\IO\Win32Marshal.cs">
       <Link>System\IO\Win32Marshal.cs</Link>
     </Compile>
-    <Compile Include="..\..\Common\src\System\SR.cs">
-      <Link>System\SR.cs</Link>
-    </Compile>
-    <Compile Include="..\..\Common\src\System\ArrayT.cs">
-      <Link>System\ArrayT.cs</Link>
-    </Compile>
+    <Compile Include="Interop\Interop.Windows.cs" />
+    <Compile Include="Microsoft\Win32\SafeHandles\SafePipeHandle.Windows.cs" />
+    <Compile Include="System\IO\Pipes\AnonymousPipeServerStream.Windows.cs" />
+    <Compile Include="System\IO\Pipes\IOCancellationHelper.Windows.cs" />
+    <Compile Include="System\IO\Pipes\NamedPipeClientStream.Windows.cs" />
+    <Compile Include="System\IO\Pipes\NamedPipeServerStream.Windows.cs" />
+    <Compile Include="System\IO\Pipes\PipeAsyncResult.cs" />
+    <Compile Include="System\IO\Pipes\PipeStream.Windows.cs" />
+    <Compile Include="System\IO\Pipes\PipeStreamAsyncResult.cs" />
   </ItemGroup>
+
+  <ItemGroup Condition=" '$(OS)' == 'Unix' ">
+    <Compile Include="..\..\Common\src\System\NotImplemented.cs"> <!-- TODO: Remove once implemented -->
+      <Link>System\NotImplemented.cs</Link>
+    </Compile>
+    <Compile Include="Microsoft\Win32\SafeHandles\SafePipeHandle.Unix.cs" />
+    <Compile Include="System\IO\Pipes\AnonymousPipeServerStream.Unix.cs" />
+    <Compile Include="System\IO\Pipes\NamedPipeClientStream.Unix.cs" />
+    <Compile Include="System\IO\Pipes\NamedPipeServerStream.Unix.cs" />
+    <Compile Include="System\IO\Pipes\PipeStream.Unix.cs" />
+  </ItemGroup>
+
   <!-- Resource files -->
   <ItemGroup>
     <EmbeddedResource Include="Resources\Strings.resx">
