+++ src/System.IO.Pipes/src/System/IO/Pipes/AnonymousPipeServerStream.Unix.cs
@@ -1,0 +1,26 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Microsoft.Win32.SafeHandles;
+using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
+using System.Security;
+
+namespace System.IO.Pipes
+{
+    /// <summary>
+    /// Anonymous pipe server stream
+    /// </summary>
+    public sealed partial class AnonymousPipeServerStream : PipeStream
+    {
+        // Creates the anonymous pipe.
+        [SecurityCritical]
+        private void Create(PipeDirection direction, HandleInheritability inheritability, int bufferSize)
+        {
+            Debug.Assert(direction != PipeDirection.InOut, "Anonymous pipe direction shouldn't be InOut");
+            Debug.Assert(bufferSize >= 0, "bufferSize is negative");
+
+            throw NotImplemented.ByDesign; // TODO: Implement this
+        }
+    }
+}
