+++ src/System.IO.Pipes/src/System/IO/Pipes/IOCancellationHelper.Windows.cs
