+++ src/System.IO.Pipes/src/System/IO/Pipes/NamedPipeClientStream.Unix.cs
@@ -1,0 +1,47 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Microsoft.Win32.SafeHandles;
+using System.Diagnostics.CodeAnalysis;
+using System.Runtime.InteropServices;
+using System.Security;
+using System.Security.Principal;
+using System.Threading;
+
+namespace System.IO.Pipes
+{
+    /// <summary>
+    /// Named pipe client. Use this to open the client end of a named pipes created with 
+    /// NamedPipeServerStream.
+    /// </summary>
+    public sealed partial class NamedPipeClientStream : PipeStream
+    {
+        private static string GetPipePath(string serverName, string pipeName)
+        {
+            throw NotImplemented.ByDesign; // TODO: Implement this
+        }
+
+        // Waits for a pipe instance to become available. This method may return before WaitForConnection is called
+        // on the server end, but WaitForConnection will not return until we have returned.  Any data writen to the
+        // pipe by us after we have connected but before the server has called WaitForConnection will be available
+        // to the server after it calls WaitForConnection. 
+        [SecurityCritical]
+        private void ConnectInternal(int timeout, CancellationToken cancellationToken, int startTime)
+        {
+            throw NotImplemented.ByDesign; // TODO: Implement this
+        }
+
+        public int NumberOfServerInstances
+        {
+            [SecurityCritical]
+            [SuppressMessage("Microsoft.Security", "CA2122:DoNotIndirectlyExposeMethodsWithLinkDemands", Justification = "Security model of pipes: demand at creation but no subsequent demands")]
+            get
+            {
+                CheckPipePropertyOperations();
+
+                throw NotImplemented.ByDesign; // TODO: Implement this
+            }
+        }
+
+    }
+}
