+++ src/System.IO.Pipes/src/System/IO/Pipes/NamedPipeServerStream.Unix.cs
@@ -1,0 +1,79 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Microsoft.Win32.SafeHandles;
+using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
+using System.Runtime.InteropServices;
+using System.Security;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace System.IO.Pipes
+{
+    /// <summary>
+    /// Named pipe server
+    /// </summary>
+    public sealed partial class NamedPipeServerStream : PipeStream
+    {
+        [SecurityCritical]
+        private void Create(string pipeName, PipeDirection direction, int maxNumberOfServerInstances,
+                PipeTransmissionMode transmissionMode, PipeOptions options, int inBufferSize, int outBufferSize,
+                HandleInheritability inheritability)
+        {
+            Debug.Assert(pipeName != null && pipeName.Length != 0, "fullPipeName is null or empty");
+            Debug.Assert(direction >= PipeDirection.In && direction <= PipeDirection.InOut, "invalid pipe direction");
+            Debug.Assert(inBufferSize >= 0, "inBufferSize is negative");
+            Debug.Assert(outBufferSize >= 0, "outBufferSize is negative");
+            Debug.Assert((maxNumberOfServerInstances >= 1 && maxNumberOfServerInstances <= 254) || (maxNumberOfServerInstances == MaxAllowedServerInstances), "maxNumberOfServerInstances is invalid");
+            Debug.Assert(transmissionMode >= PipeTransmissionMode.Byte && transmissionMode <= PipeTransmissionMode.Message, "transmissionMode is out of range");
+
+            throw NotImplemented.ByDesign; // TODO: Implement this
+        }
+
+        // This will wait until the client calls Connect().  If we return from this method, we guarantee that
+        // the client has returned from its Connect call.   The client may have done so before this method 
+        // was called (but not before this server is been created, or, if we were servicing another client, 
+        // not before we called Disconnect), in which case, there may be some buffer already in the pipe waiting
+        // for us to read.  See NamedPipeClientStream.Connect for more information.
+        [SecurityCritical]
+        [SuppressMessage("Microsoft.Security", "CA2122:DoNotIndirectlyExposeMethodsWithLinkDemands", Justification = "Security model of pipes: demand at creation but no subsequent demands")]
+        public void WaitForConnection()
+        {
+            CheckConnectOperationsServer();
+
+            throw NotImplemented.ByDesign; // TODO: Implement this
+        }
+
+        public Task WaitForConnectionAsync(CancellationToken cancellationToken)
+        {
+            if (cancellationToken.IsCancellationRequested)
+            {
+                return TaskHelpers.FromCancellation(cancellationToken);
+            }
+
+            return Task.Factory.StartNew(s => ((NamedPipeServerStream)s).WaitForConnection(),
+                this, cancellationToken, TaskCreationOptions.DenyChildAttach, TaskScheduler.Default);
+        }
+
+        [SecurityCritical]
+        public void Disconnect()
+        {
+            CheckDisconnectOperations();
+
+            throw NotImplemented.ByDesign; // TODO: Implement this
+        }
+
+        // Gets the username of the connected client.  Not that we will not have access to the client's 
+        // username until it has written at least once to the pipe (and has set its impersonationLevel 
+        // argument appropriately). 
+        [SecurityCritical]
+        public String GetImpersonationUserName()
+        {
+            CheckWriteOperations();
+
+            throw NotImplemented.ByDesign; // TODO: Implement this
+        }
+    }
+}
