+++ src/System.IO.Pipes/src/System/IO/Pipes/PipeDirection.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace System.IO.Pipes
+{
+    public enum PipeDirection
+    {
+        In = 1,
+        Out = 2,
+        InOut = In | Out,
+    }
+}
