+++ src/System.IO.Pipes/src/System/IO/Pipes/PipeStream.Unix.cs
@@ -1,0 +1,175 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Microsoft.Win32.SafeHandles;
+using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
+using System.Diagnostics.Contracts;
+using System.Runtime.InteropServices;
+using System.Security;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace System.IO.Pipes
+{
+    public abstract partial class PipeStream : Stream
+    {
+        /// <summary>Throws an exception if the supplied handle does not represent a valid pipe.</summary>
+        /// <param name="safePipeHandle">The handle to validate.</param>
+        internal static void ValidateHandleIsPipe(SafePipeHandle safePipeHandle)
+        {
+            throw NotImplemented.ByDesign; // TODO: Implement this
+        }
+
+        /// <summary>Initializes the handle to be used asynchronously.</summary>
+        /// <param name="handle">The handle.</param>
+        [SecurityCritical]
+        private void InitializeAsyncHandle(SafePipeHandle handle)
+        {
+            // nop
+        }
+
+        [SecurityCritical]
+        private unsafe int ReadCore(byte[] buffer, int offset, int count)
+        {
+            Debug.Assert(_handle != null, "_handle is null");
+            Debug.Assert(!_handle.IsClosed, "_handle is closed");
+            Debug.Assert(CanRead, "can't read");
+            Debug.Assert(buffer != null, "buffer is null");
+            Debug.Assert(offset >= 0, "offset is negative");
+            Debug.Assert(count >= 0, "count is negative");
+
+            throw NotImplemented.ByDesign; // TODO: Implement this
+        }
+
+        [SecurityCritical]
+        private unsafe void WriteCore(byte[] buffer, int offset, int count)
+        {
+            Debug.Assert(_handle != null, "_handle is null");
+            Debug.Assert(!_handle.IsClosed, "_handle is closed");
+            Debug.Assert(CanWrite, "can't write");
+            Debug.Assert(buffer != null, "buffer is null");
+            Debug.Assert(offset >= 0, "offset is negative");
+            Debug.Assert(count >= 0, "count is negative");
+
+            throw NotImplemented.ByDesign; // TODO: Implement this
+        }
+
+        // Blocks until the other end of the pipe has read in all written buffer.
+        [SecurityCritical]
+        public void WaitForPipeDrain()
+        {
+            CheckWriteOperations();
+            if (!CanWrite)
+            {
+                throw __Error.GetWriteNotSupported();
+            }
+
+            throw NotImplemented.ByDesign; // TODO: Implement this
+        }
+
+        // ********************** Public Properties *********************** //
+
+        // Gets the transmission mode for the pipe.  This is virtual so that subclassing types can 
+        // override this in cases where only one mode is legal (such as anonymous pipes)
+        public virtual PipeTransmissionMode TransmissionMode
+        {
+            [SecurityCritical]
+            [SuppressMessage("Microsoft.Security", "CA2122:DoNotIndirectlyExposeMethodsWithLinkDemands", Justification = "Security model of pipes: demand at creation but no subsequent demands")]
+            get
+            {
+                CheckPipePropertyOperations();
+
+                if (_isFromExistingHandle)
+                {
+                    throw NotImplemented.ByDesign; // TODO: Implement this
+                }
+                else
+                {
+                    return _transmissionMode;
+                }
+            }
+        }
+
+        // Gets the buffer size in the inbound direction for the pipe. This checks if pipe has read
+        // access. If that passes, call to GetNamedPipeInfo will succeed.
+        public virtual int InBufferSize
+        {
+            [SecurityCritical]
+            [SuppressMessage("Microsoft.Security", "CA2122:DoNotIndirectlyExposeMethodsWithLinkDemands")]
+            get
+            {
+                CheckPipePropertyOperations();
+                if (!CanRead)
+                {
+                    throw new NotSupportedException(SR.NotSupported_UnreadableStream);
+                }
+
+                throw NotImplemented.ByDesign; // TODO: Implement this
+            }
+        }
+
+        // Gets the buffer size in the outbound direction for the pipe. This uses cached version 
+        // if it's an outbound only pipe because GetNamedPipeInfo requires read access to the pipe.
+        // However, returning cached is good fallback, especially if user specified a value in 
+        // the ctor.
+        public virtual int OutBufferSize
+        {
+            [SecurityCritical]
+            [SuppressMessage("Microsoft.Security", "CA2122:DoNotIndirectlyExposeMethodsWithLinkDemands", Justification = "Security model of pipes: demand at creation but no subsequent demands")]
+            get
+            {
+                CheckPipePropertyOperations();
+                if (!CanWrite)
+                {
+                    throw new NotSupportedException(SR.NotSupported_UnwritableStream);
+                }
+
+                int outBufferSize;
+
+                // Use cached value if direction is out; otherwise get fresh version
+                if (_pipeDirection == PipeDirection.Out)
+                {
+                    outBufferSize = _outBufferSize;
+                }
+                else
+                {
+                    throw NotImplemented.ByDesign; // TODO: Implement this
+                }
+
+                return outBufferSize;
+            }
+        }
+
+        public virtual PipeTransmissionMode ReadMode
+        {
+            [SecurityCritical]
+            get
+            {
+                CheckPipePropertyOperations();
+
+                // get fresh value if it could be stale
+                if (_isFromExistingHandle || IsHandleExposed)
+                {
+                    throw NotImplemented.ByDesign;
+                }
+                return _readMode;
+            }
+            [SecurityCritical]
+            [SuppressMessage("Microsoft.Security", "CA2122:DoNotIndirectlyExposeMethodsWithLinkDemands", Justification = "Security model of pipes: demand at creation but no subsequent demands")]
+            set
+            {
+                CheckPipePropertyOperations();
+                if (value < PipeTransmissionMode.Byte || value > PipeTransmissionMode.Message)
+                {
+                    throw new ArgumentOutOfRangeException("value", SR.ArgumentOutOfRange_TransmissionModeByteOrMsg);
+                }
+
+                throw NotImplemented.ByDesign; // TODO: Implement this
+            }
+        }
+
+    }
+}
+
+
