+++ src/System.IO.Pipes/src/System/IO/Pipes/PipeTransmissionMode.cs
@@ -1,0 +1,11 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace System.IO.Pipes
+{
+    public enum PipeTransmissionMode
+    {
+        Byte = 0,
+        Message = 1,
+    }
+}
